XiFluid: Added complete slip-velocity term to the unburnt gas energy equation
In the current formulation for inhomogeneous combustion this additional term
cannot be applied for consistency reasons so it is only applied for homogeneous
combustion.  A new formulation for inhomogeneous combustion in being developed
to support the slip-velocity term.

The corresponding burnt gas slip-velocity term is included in the code but
commented-out until it is required for the solution of burnt gas properties.