extrudeModel: Improve usability of input
The models now don't read nLayers and expansionRatio unless they are
needed, and these values no longer default to one for models that do.
The coeffs sub-dictionary now always corresponds to the model name. A
number of examples and tutorials have been updated and a couple of the
tutorials have also been corrected so that they correctly create
multiple layers rather than defaulting to one.