Lagrangian: More injection velocity conditions
A number of additional options have been added for defining the velocity
of particles injected at a point or in a disk. These are:

'fanVelocity' provides a fan velocity profile, characterised by a
centreline velocity, a normal, and inner and outer fan angles. The
centreline velocity and normal are Function1s and can vary with time.
This is likely to be used in conjunction with a point injection model.
Example:

    <LagrangianModelName>
    {
        type            fanVelocity;
        Ucentre
        {
            type            table;
            values
            (
                (0 (0.1 0 0))
                (0.5 (0.5 0 0))
            );
        }
        thetaInner      15 [deg];
        thetaOuter      22.5 [deg];
    }

'flowRateConeDiskVelocity' provides a conical velocity profile over a
disk, characterised by inner and outer cone angles. The velocity
magnitude is obtained from the area of the disk and the flow rate of the
corresponding flowRateNumber condition. This condition must be used in
conjunction with a disk injection model and a flow rate condition from
the number. Example:

    <LagrangianModelName>
    {
        type            flowRateConeDiskVelocity;
        thetaInner      5 [deg];
        thetaOuter      30 [deg];
    }

'totalPressureConeVelocity' provides a conical velocity profile,
characterised by inner and outer cone angles, with a magnitude
calculated from the difference between a specified upstream pressure and
the local interpolated pressure.

    <LagrangianModelName>
    {
        type            totalPressureConeVelocity;
        direction       (1 0 0);
        thetaInner      15 [deg];
        thetaOuter      22.5 [deg];
        p0              1 [bar];
    }

The functionality components of these conditions (i.e., cone, cone-disk,
fan, total pressure, flow-rate) have been separated into reusable mix-in
classes. This means further combinations (e.g.,
totalPressureFanVelocity) can be conveniently created in future as
needed.