hexRef8: Is now a PermanentMeshObject held by the polyMesh
This ensures that there is a single unique instance of the hexRef8 class holding
the refinement state of the mesh and class like fvMeshTopoChangers::refiner
lookup the hexRef8 rather than creating a local instance.  This means that
fvMeshTopoChangers::list now supports more than one fvMeshTopoChangers::refiner,
each of which can refine the mesh in different ways but all using the same
hexRef8 to ensure consistency.