functionObjects: Remove cached references to solver-dependent data
Anything solver-dependent should be looked up at the point of usage.
This is required by foamPostProcess which constructs the solver afresh
at each execution time. The solver is replaced in this way in order to
ensure that all properties are re-read and initialised in exactly the
same way as for a restart.