interpolation: Added interpolation of the gradient
Scalar and vector interpolation classes now have interpolateGrad methods
that return the gradient of the interpolated field. Note that this is
intended to be a direct gradient of the interpolating function, and not
an interpolation of a finite-volume gradient of the field. The latter
could (and should) be constructed manually if needed.

Only the cellPoint method (and it's derivatives) currently provide an
implementation of gradient interpolation; other methods will throw a
run-time error. In general, any interpolation method that has an
interpolating function with a valid first gradient could support
gradient interpolation. So, cell interpolation will never support it,
but all the other methods should eventually be able to.