XiFluid::thermophysicalPredictor: Removed absoluteInternalEnergy option
It is not possible to construct a robust and reliable pressure-work term for the
unburnt internal energy equation from the mean flux and modelled unburnt-burnt
gas velocity difference, the numerical error in the work estimate is too great,
it is much more reliable to use an enthalpy formulation for which the
pressure-work it obtained from the rate of change of pressure which is
consistent with the basic modelling assumption that the unburnt and burnt gas
pressures are the same.  For consistency the mean energy and unburnt energy
forms must be the same so both are now based on the absolute enthalpy and the
internal energy option has been removed.