Lagrangian::turbulentDispersion: Prevent excessive iteration
This change comprises two fixes. Firstly, the turbulence has been
constrained to zero for particles that are in contact with walls. This
prevents a turbulent velocity fluctuation from being generated that
points into the wall. This in turn prevents a drag force into the wall
that causes the particle to vibrate in conjunction with a rebound
velocity boundary condition.

Secondly, when the turbulent time scale becomes small the model now
stops treating eddies individually and instead applies a statistical
estimate of their combined effect. This prevents excessive iteration
near walls when the eddies become very small and the number of eddies
increases significantly. The number of eddies at which this happens can
be controlled with a new 'maxDiscreteEddies' switch, which is set to 32
by default.