Lagrangian: Mass transfers and coupling to multiphase
Mass transfer modelling has been added to field-based Lagrangian. Two
models are included at present:

The 'constantFluxCarrierTransfer' model represents mass transfer into
the carrier phase with a constant surface volume or mass flux. This
model provides a crude approximation of phase change which neglects any
thermal effects. It can be used by clouds and in solvers that do not
feature any thermodynamic modelling. Example:

    evaporation
    {
        type        constantFluxCarrierTransfer;
        volumeFlux  1 [ml/cm^2/s];
    }

The 'collisionPhaseTransfer' model represents the absorption of droplets
or bubbles into a corresponding Eulerian phase as a result of collisions
with that phase (and potentially other non-carrier Eulerian phases).
This can be used to represent (e.g.) a spray of droplets hitting and
being absorbed by a VoF interface. The collision probability is
calculated as proportional to the Eulerian phase's surface area per unit
volume and the distance travelled by the particle. Example:

    collision
    {
        type        collisionPhaseTransfer;
    }

Multiphase coupling has also been added to field-based Lagrangian. The
Eulerian phases with which the cloud interacts can be identified within
the cloud fvModel or functionObject specification. E.g., in
constant/fvModels:

    droplets
    {
        type        parcelCloud;
        libs        ("libLagrangianCloud.so");

        // The corresponding Eulerian phase. Used as a source for
        // physical property information. Also used as the phase to
        // which volume or mass is transferred to as a result of the
        // collisionPhaseTransfer model.
        phase       water;

        // The carrier Eulerian phase. The phase that surrounds the
        // cloud's particles and with which the cloud exchanges
        // momentum, heat, etc... The phase with which volume or mass is
        // transferred to as a result of the constantFluxCarrierTransfer
        // model.
        carrierPhase air;
    }

These 'phase' and 'carrierPhase' settings are optional. If they are
omitted then the cloud will be considered to be of an entirely different
phase and the phase mixture will be considered the carrier.