laminarModels::Maxwell,Giesekus,PTT: Added support for modal viscosity specification
When multiple modes are specified both the time-scale and non-Newtonian
viscosity is are now required, providing greater flexibility.  Additionally the
optional modes entry in momentumTransport is now a sub-dictionary rather than a
list of dictionaries, with each mode given a name to make it easier to
manipulate the entries with internal substitutions or external changes using
foamDictionary, e.g.:

    modes
    {
        mode0
        {
            nuM             0.01476;
            lambda          0.018225;
        }

        mode1
        {
            nuM             0.02476;
            lambda          0.028225;
        }
    }

Alternatively for cases where a single mode is sufficient the previous syntax

    nuM             0.01476;
    lambda          0.018225;

is still supported.