fvMesh, polyMesh: Write geometry fields with additional precision
Moved points, old-time mesh volumes and mesh fluxes are now all written
in ascii format at a higher level of precision, similar to the
additional levels of precision specified by various utilties. In
addition, the additional levels of precision have been named and added
to IOstream and now relate to the precision of the floating point type.

Resolves bug report https://bugs.openfoam.org/view.php?id=4273