fvMesh: Only write mesh fluxes when necessary
When using a first order time-scheme only the mesh fluxes relating to
the current time-step are needed, so it is not necessary write the mesh
fluxes as on restart the values loaded from disk will relate to the
previous timestep. Writing of the mesh fluxes has been deactivated in
this case in order to save disk space and compute time. Mesh fluxes are
still written when using a second-order time scheme.