fvModels: reactionDriven: Fix when both phases are reacting
This model now only causes phase change as a result of reactions in the
first phase. This means the transferred species can undergo further
reaction in the second phase without the model transferring them back to
the first. If reaction-driven transfer is needed in both directions (for
different sets of species) then two models should be defined by the
user, with the phases in opposite orders.