XiFluid: New solution algorithm and two-phase thermodynamics
The new solution algorithm for the Weller b-Xi combustion model in XiFluid is
based on a complete and consistent two-phase formulation in which the unburnt
and burnt gas enthalpy equations are solved in conservative form and the mean
properties calculated by averaging the corresponding unburnt and burnt gas
properties.  This avoids the need for a mean enthalpy equation and the unity
Lewis number assumption fundamental to that approximation.  The disadvantage of
solving for the unburnt and burnt gas properties is these phase equations must
be stabilised so that they can be solved everywhere in the domain, even where
the corresponding phase-fraction -> 0 which required the addition of
stabilisation transport terms which -> 0 where the phase-fraction is larger than
a specified lower limit, which defaults to 1e-3.

Additionally the stability of this fully conservative two-phase approach is
critically dependent on the boundedness and conservation of the regress variable
b in order to provide consistent fluxes of b, the flame propagation rate and
unburnt gas consumption rate.  These constraints are fulfilled by solving for b
utilising the latest MULES limiter.  Firstly a 1st-order fully implicit bounded
conservative formulation of the b transport equation is solved, including
convection, diffusion and flame-propagation terms and then an explicit
higher-order correction equation is formulated to which the MULES limiter is
applied to ensure that the corrected b solution remains bounded.

The unburnt and burnt gas energy equations are solved for the sensible rather
than absolute enthalpy so that the thermal and specie diffusive transport are
decoupled avoiding the unity Lewis number limitation of the previous
formulation.  For numerical reasons it is preferable to solve for enthalpy
rather than internal energy because the pressure time derivative needed for the
pressure work term is more accurate than the unburnt or burnt gas dilatation
term required by the internal energy equation pressure work due to the single
pressure closure used in this formulation.

Simple diffusive turbulent transport closure for the unburnt and burnt gas
energy and specie equations is currently hard-coded with unity Prandtl and
Schmidt numbers for backwards-compatibility but there is no formal restriction
on this closure modelling and in the future it will be abstracted and made
run-time selectable as it is in other solvers.

The specification of the fvSchemes is simpler than in the previous solver as the
multivariate consistency of transport required by the unity Lewis number
constraint is no longer necessary so the schemes for each of the transport terms
are listed separately, e.g. for the transport of b and Xi:

    div(phi,b)          Gauss vanLeer;
    div(phiXi,Xi)       Gauss vanLeer;

and for the unburnt gas enthalpy:

    div(phib,h.u)       Gauss limitedLinear 1;
    div(phib,K)         Gauss limitedLinear 1;

where phib is the flux of b

and for the burnt gas enthalpy:

    div(phic,h.b)       Gauss limitedLinear 1;
    div(phic,K)         Gauss limitedLinear 1;

where phic is the flux of c, the regress variable (1 - b).

Now that the thermophysical implementation in XiFluid is two-phase the input
files follow the multiphase naming rules of the VoF and Euler-Euler solvers with
a phase-name post-fix, 'u' for unburnt and 'b' for burnt.  Thus the physical
properties of the unburnt and burnt gases are specified in the
constant/physicalProperties.u and constant/physicalProperties.b files and the
unburnt and burnt gas temperature fields named T.u and T.b respectively (there
is no need to provide a mean temperature field T).

For homogeneous combustion the new ubPsiThermo unburnt or burnt gas
thermodynamics type is used with the uHomogeneousMixture mixture type, specified
in physicalProperties.u e.g.

thermoType
{
    type            ubPsiThermo;
    mixture         uHomogeneousMixture;
    transport       const;
    thermo          hConst;
    equationOfState perfectGas;
    specie          specie;
    energy          sensibleEnthalpy;
}

and the bHomogeneousMixture mixture type is specified in physicalProperties.b
e.g.

thermoType
{
    type            ubPsiThermo;
    mixture         bHomogeneousMixture;
    transport       const;
    thermo          hConst;
    equationOfState perfectGas;
    specie          specie;
    energy          sensibleEnthalpy;
}

For inhomogeneous combustion the uInhomogeneousMixture and bInhomogeneousMixture
mixture types are used and the unburnt gas fuel mass fraction fu.u required.  If
the mixture is rich the burnt gas total fuel mass fraction ft.b (commonly known
as the mixture fraction) is also required.  With exhaust gas recirculation (EGR)
the mass-fraction of products in the unburnt gas egr.u is required but no
additional composition specification is needed for the burnt gas.

Dynamic compilation of ubPsiThermo with unusual combinations of thermophysical
models is supported but only with sensibleEnthalpy as the energy type as
explained above.

The specification of fvModels, fvConstraints and functionObjects relating to the
now two-phase thermodynamics must specify the phase they relate to, e.g. in the
engine2Valve2D tutorial case the unburnt and burnt gas temperatures are limited
to stabilise the simulation during valve opening and closing by specifying the
following constraints in constant/fvConstraints:

    #includeConstraint limitTemperature(phase=u, min=290, max=3000)
    #includeConstraint limitTemperature(phase=b, min=290, max=3000)

All the tutorial cases:
    1D  engine2Valve2D  kivaTest  moriyoshiHomogeneous  stratified
have been updated for the new version of XiFluid.