flowType function object: rewritten to visualise vortices effectively.
flowType alculates and writes a field in the range -1 to 1, in which:
-  1 = rotational flow
-  0 = parallel shear (50% rotational, 50% shear)
- -1 = pure shear deformation

The field is calculated as:
flowType = -(devGradU && devGradU.T())/magSqr(devGradU)

where devGradU = dev(fvc::grad(U)).  The flowType parameter is designed
for visualisation of vortices, by extracting an iso-surface of these
fields at a specified value.  The parameter is similar to Q and Lambda2,
except that Q and Lambda2 represent the rotational speed, or "strength",
of vortices whereas flowType is normalised to a scale between -1 and 1,
making it independent of vortex strength.

The normalisation of flowType provides the advantage that it is much
easier to select a value of iso-surface to visualise vortices
effectively.  Generally a value of 0.2 (or between 0.05 - 0.2) is
recommended.  With Q and Lambda2 the choice of an effective iso-surface
value depends on the strength of vortices for the specific CFD problem.

The flowType function object writes an additional magSqrDevGradU field
corresponding to the denominator in the flowType field.  An iso-surface
can then be coloured by the magSqrDevGradU field to indicate the
strength of the vortices as they move with the flow.