XiFluid: Abstracted the unburnt and burnt gas thermophysical transport modelling
The unburnt and burnt gas thermal and specie transport is now handled by the
general PhaseThermophysicalTransportModel used for multiphaseEuler and other
multiphase solvers providing support for non-unity Lewis number transport,
Fickian and even Maxwell-Stefan.

The use of this general abstraction required a number of structural changes to
XiFluid in order to make it compatible with other multiphase solvers, in
particular the new RASModels::ubMomentumTransportModel class is used to wrap the
mixture RAS model to provide the phase turbulence properties to
PhaseThermophysicalTransportModel which required the phase densities to be
cached requiring the unburnt and burnt gas thermo to change from ubPsiThermo to
ubRhoThermo.

The CHT ubCoupledTemperatureFvPatchScalarField boundary condition has been
update to obtain kappaEff for the unburnt and burnt gases directly from the
corresponding PhaseThermophysicalTransportModel.