etc/controlDict -> etc/configDict: Renamed for consistency with the new system/configDict
The global configuration controls which were in etc/controlDict and optionally
overridden by corresponding settings in the case system/controlDict are now in
etc/configDict and optionally in system/configDict to separate them from
system/controlDict to allow the latter to support any of the dictionary
functionEntries which were not supported previously due to parsing restrictions
during the construction of global data.  Now parsing restrictions are limited to
etc/configDict and the optional system/configDict.