GAMGProcAgglomeration: Rationalised naming and added documentation to encourage testing
The most promising processor agglomeration option is pair-wise:

Description
    Pair-wise processor agglomerator.

    Neighbouring pairs of processors are agglomerated when the number of
    agglomerated cells per processor in the level is below \c
    minCellsPerProcessor and the pairs chosen based on the largest number of
    connecting processor faces.

Usage
    Example of GAMG solver settings with pair-wise processor agglomeration
    \verbatim
    p
    {
        solver          GAMG;
        smoother        GaussSeidel;

        processorAgglomeration
        {
            agglomerator         pair;
            minCellsPerProcessor 300;
        }

        tolerance       1e-8;
        relTol          0;
    }
    \endverbatim

On large numbers of processors it is advantageous to reduce minCellsPerProcessor
to 1 with which the cells and processors agglomerate until the minimum number of
cells at the coarsest level is equal to the number of processors.  Unfortunately
it is not currently possible to reduce this further because processor
agglomeration takes place after all the cell agglomeration rather than occurring
after the cell agglomeration at each level.  Further development is required to
make this functionality more generally useful and effective.