dictionary: New code compilation optimisation using #codeBlock ... #endCodeBlock
Each #calc and #codeStream in a dictionary generates a coded function in a file
which is compiled, linked and executed to inject results into the dictionary
entries, for example the blockMeshDict in tutorials/incompressibleFluid/venturiTube
    .
    .
    .
    radIn      #calc "$diameter / 2.0";
    boxIn      #calc "3.0*$radIn / 10.0";

    radVen     #calc "$radIn / 2.0";
    boxVen     #calc "$boxIn / 2.0";

    xInSt      0;
    xInEnd     $diameter;
    xVenSt     #calc "$xInEnd  + 1.25*$diameter";
    xVenEnd    #calc "$xVenSt  + 0.5*$diameter";
    xOutSt     #calc "$xVenEnd + 2.5*$diameter";
    xOutEnd    #calc "$xOutSt  + $diameter";

generates and compiles 8 file which  takes ~6s to generate the mesh.
With the new #codeBlock ... #endCodeBlock structure
    .
    .
    .
    #codeBlock

    radIn      #calc "$diameter / 2.0";
    boxIn      #calc "3.0*$<scalar>radIn / 10.0";

    radVen     #calc "$<scalar>radIn / 2.0";
    boxVen     #calc "$<scalar>boxIn / 2.0";

    xInSt      0;
    xInEnd     $diameter;
    xVenSt     #calc "$<scalar>xInEnd  + 1.25*$diameter";
    xVenEnd    #calc "$<scalar>xVenSt  + 0.5*$diameter";
    xOutSt     #calc "$<scalar>xVenEnd + 2.5*$diameter";
    xOutEnd    #calc "$<scalar>xOutSt  + $diameter";

    #endCodeBlock

the 8 function created by the #calc entries are collated into a single file
which is compiled and linked to provide the 8 functions and this reduces the
mesh generation time ~1s.

Note that in the above example to make the 8 #calc entries independent and
possible to include in a single file each lookup must have the type provided (in
this case <scalar>) because the type of the result of previous evaluation is not
know at the point of generation of the code of the dependent #calc.

A more complex example is in test/dictionary/testCalc which has been updated to
include all the entries within a single #codeBlock ... #endCodeBlock reducing
the compilation and run time from 13s to 1.7s.

There are however a few issues/limitations with the implementation but
fundamental to the approach and relating to scope and interaction between
dictionary entries, in particular within the #codeBlock...#endCodeBlock:
  * only #calc, #codeStream, #calcInclude and #neg functionEntries are supported
  * #calcInclude should be used for #calc AND #codeStream rather than the
    codeInclude available with #codeStream
  * The #calc and #codeStream entries within #codeBlock ... #endCodeBlock must
    be independent requiring type specification of values to be looked-up which
    result from other code evaluations within the block as shown in the example
    above.