readUpdate: Optimise sequential updates
This change prevents readUpdate from scanning back all the way to time
zero in most cases by limiting the minimum time considered to that of
the previous update. This only applies to updates that change to a later
time. If the change moves to an earlier time then such an optimisation
is not possible, and the scanning will extend to time zero as before.

This optimisation prevents the gradual slowing down of certain pre- and
post-processing operations observed when operating on cases with a large
number of time directories.