populationBalance: Simplify control of coalescence and breakup
Coalescence and breakup models are no longer provided as lists. It is
not meaningful to use multiple such models simultaneously without also
specifying how they are to be combined/blended (e.g., 30% of one, 70% of
the other). If such combinations are needed in future this could be
achieved by having a model that wraps other models.

The removal of the lists of models also means that dictionary
substitution can be used conveniently within the model specifications.

There is now no distinction from the users perspective between breakup
and binary breakup models. They are all just selected as breakup models.

Like diameter and shape models, coalescence and breakup models can be
specified as a dictionary (preferred if there are coefficients) or as a
type name with a separate optional '<type>Coeffs' dictionary (preferred
if there are no coefficients and the '<type>Coeffs' dictionary can be
omitted).

An example of the new syntax is as follows:

    bubbles
    {
        continuousPhase liquid;

        sphericalDiameters
        {
            type            exponential;
            min             1e-4;
            max             1.14e-2;
        }

        shapeModel      spherical;

        coalescenceModel
        {
            type            PrinceBlanch;
            turbulence      on;
            buoyancy        on;
            laminarShear    off;
            C1              0.05;
        }

        breakupModel    LehrMilliesMewes;
    }